wyspp/requirements.txt
asgiref==3.3.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
defusedxml==0.7.0rc1
Django==3.1.4
django-braces==1.14.0
django-cors-headers==3.6.0
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
fcm-django==0.3.7
google-api-core==1.24.0
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-storage==1.34.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
idna==2.10
mysqlclient==2.0.2
oauthlib==3.1.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.9
PyJWT==1.7.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
urllib3==1.26.2
